create folder for sportradar data types and rename existing type file
