+++ app/types/sportradar/game-summary.ts
